Finalized Google Login (OAuth2)
I have reinimplemented nickname for confidentiality purpose (More on this in rapport). I have made it so that nickname is integrated for external service login. We removed nickname when attempting to implement google login since it dit not quit work with our already existing setup. Now i have fixed that.